Added ServerFacade register function.

+++ shared/src/main/java/exception/ResponseException.java
@@ -15,7 +15,7 @@ public class ResponseException extends Exception {
 
     public static ResponseException fromJson(String json) {
         var map = new Gson().fromJson(json, HashMap.class);
-        int status = (int) map.get("status");
+        int status = ((Double) map.get("status")).intValue();
         String message = map.get("message").toString();
         return new ResponseException(message, status);
     }
